✨ split exports and install
