+++ src/index.js
@@ -19,25 +19,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-import * as NcComponents from './components/index.js'
 
-/**
- * @param {object} Vue The vue instance
- */
-function install(Vue) {
-	Object.values(NcComponents).forEach((component) => {
-		Vue.component(component.name, component)
-	})
-}
-
-if (typeof window !== 'undefined' && window.Vue) {
-	install(window.Vue)
-}
-
-export default {
-	install,
-	...NcComponents,
-}
 export * from './components/index.js'
 export * from './directives/index.js'
 export * from './mixins/index.js'
